requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv==1.2.1
aiohttp==3.13.3
aiosqlite==0.22.1
